Made it even shorter somehow
I was going to make a repo for the YAZ assignment and noticed you forked my assignment so I went to check for comments and remembered that I thought of a few ways I could make it shorter and then I spent my lunch break working on it. In the spec it says it took them 88 lines of code to write the full game so I decided to challenge myself to do less but I realized that wouldn't be that hard so I decided to make it as short as possible without doing something dumb like putting all the code in main. Just so we're clear this code was not meant to be readable and it was not meant to be good (maybe clever) just as short as possible without using a dumb trick.